columns
This patch adds an extra layout to dwm called `col` in which the windows in the
master area are arranged in colums of equal size. The number of columns is
always nmaster + 1, and the last column is a stack of leftover windows just
like the normal tile layout. It effectively acts like the default tiling mode,
except provides for vertical instead of horizontal master windows.

Co-authored-by: Evan Gates (emg) <evan.gates@gmail.com>
Co-authored-by: Noah Rosser <noah.rosser@gmail.com>